requirements.txt
python-dotenv~=1.0.1
Werkzeug~=3.1.3
SQLAlchemy~=2.0.37
Flask-RESTful~=0.3.10
Flask~=3.1.0
Flask-JWT-Extended~=4.7.1
Flask-Cors~=5.0.0
pandas~=2.2.3
numpy~=2.2.2